# new-Health/src/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Health Microservice deals with ECG, BP, Sugar, Stress and other Parameters",
        "version": "0.0.1",
        "title": "Health Microservice"
    },
    "host": "localhost:1000",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    },
    "paths": {
...
        "/reportissue/upload": {
            "x-swagger-router-controller": "labsdk",
            "post": {
                "tags": [
                    "report issue"
                ],
                "operationId": "reportissue",
                "description": "Upload the pdf to server with patient id",
                "parameters": [
                    {
                        "name": "pdf",
                        "in": "formData",
                        "description": "Pdf report",
                        "type": "file"
                    },
                    {
                        "name": "deviceId",
                        "in": "formData",
                        "required": true,
                        "description": "Device ID",
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "username"
                    },
                    {
                        "name": "reportType",
                        "in": "formData",
                        "required": true,
                        "description": "type of report eg BP,ECG,stress,temp",
                        "type": "string"
                    },
                    {
                        "name": "reportid",
                        "in": "formData",
                        "description": "id",
                        "type": "string"
                    },
                    {
                        "name": "firmwareVersion",
                        "in": "formData",
                        "description": "Firmware version",
                        "type": "string"
                    },
                    {
                        "name": "appVersion",
                        "in": "formData",
                        "description": "App version",
                        "type": "string"
                    },
                    {
                        "name": "phoneModel",
                        "in": "formData",
                        "description": "Phone model",
                        "type": "string"
                    },
                    {
                        "name": "osVersion",
                        "in": "formData",
                        "description": "OS Version",
                        "type": "string"
                    },
                    {
                        "name": "createdTs",
                        "in": "formData",
                        "description": "Reporting date",
                        "type": "string"
                    },
                    {
                        "name": "comment",
                        "in": "formData",
                        "description": "Feedback from customer if any",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdk/upload/{deviceId}/{patientId}/{labusername}": {
            "x-swagger-router-controller": "labsdk",
            "post": {
                "tags": [
                    "Lab SDK"
                ],
                "operationId": "uploadpdf",
                "description": "Upload the pdf to server with patient id",
                "parameters": [
                    {
                        "name": "ecgpdf",
                        "in": "formData",
                        "description": "The username of the person who want's ecg",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "deviceId",
                        "in": "path",
                        "description": "deviceId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "patient Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "labusername",
                        "in": "path",
                        "description": "labusername",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdk/fetch/{deviceId}/{patientId}": {
            "x-swagger-router-controller": "labsdk",
            "get": {
                "tags": [
                    "Lab SDK"
                ],
                "operationId": "fetchBlob",
                "description": "Upload the pdf to server with patient id",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "path",
                        "description": "deviceId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "patient Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdk/fetchlab/{deviceId}/{labusername}": {
            "x-swagger-router-controller": "labsdk",
            "get": {
                "tags": [
                    "Lab SDK"
                ],
                "operationId": "fetchreport",
                "description": "Upload the pdf to server with patient id",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "path",
                        "description": "deviceId",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "labusername",
                        "in": "path",
                        "description": "patient Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sdk/delete/blob": {
            "x-swagger-router-controller": "labsdk",
            "post": {
                "tags": [
                    "Lab SDK"
                ],
                "operationId": "deleteBlob",
                "description": "Upload the pdf to server with patient id",
                "parameters": [
                    {
                        "name": "blobUrl",
                        "in": "body",
                        "description": "deviceId",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/blob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "blob": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "ErrorResponse": {
            "title": "error",
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "title": {
                                "type": "string"
                            },
                            "code": {
                                "type": "string"
                            },
                            "detail": {
                                "type": "string"
                            },
                            "source": {
                                "type": "object",
                                "properties": {
                                    "pointer": {
                                        "type": "string"
                                    },
                                    "parameter": {
                                        "type": "string"
                                    }
                                }
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "required": [
                            "title",
                            "code",
                            "detail"
                        ]
                    }
                }
            }
        }
    }
}